+++ packages/editor/src/components/index.js
@@ -27,6 +27,7 @@ export { default as PluginDocumentSettingPanel } from './plugin-document-setting
 export { default as PluginBlockSettingsMenuItem } from './block-settings-menu/plugin-block-settings-menu-item';
 export { default as PluginMoreMenuItem } from './plugin-more-menu-item';
 export { default as PluginPostPublishPanel } from './plugin-post-publish-panel';
+export { default as PluginPostRevisionInfo } from './plugin-post-revision-info';
 export { default as PluginPostStatusInfo } from './plugin-post-status-info';
 export { default as PluginPrePublishPanel } from './plugin-pre-publish-panel';
 export { default as PluginPreviewMenuItem } from './plugin-preview-menu-item';